+++ oshi-json/src/test/java/oshi/json/software/os/OperatingSystemTest.java
@@ -54,6 +54,7 @@ public class OperatingSystemTest {
 
         assertTrue(os.getProcessCount() >= 1);
         assertTrue(os.getThreadCount() >= 1);
+        assertTrue(os.getBitness() == 32 || os.getBitness() == 64);
         assertTrue(os.getProcessId() > 0);
 
         assertTrue(os.getProcesses(0, null).length > 0);